archives/python_legacy/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1
motor
pydantic-settings
